+++ src/amun/common/histories.h
@@ -99,7 +99,7 @@ public:
   void Add(const God &god);
   Hypotheses GetSurvivors() const;
 
-  std::vector<size_t> Hypo2Batch() const;
+  std::vector<unsigned> Hypo2Batch() const;
 
   void OutputAll(const God &god);
